+++ src/@types/styled.d.ts
@@ -2,6 +2,6 @@ import "styled-components";
 import themes from "@Global/themes";
 
 declare module "styled-components" {
-  type ThemeType = typeof theme;
+  type ThemeType = typeof themes;
   export interface DefaultTheme extends ThemeType {}
 }